refactor : TimeFormat Util함수 분리
